pulling in toms changes
